requirements.txt
requests
python-dotenv
markdown
qrcode[pil]
